+++ reflective/stack_entry.hpp
@@ -42,14 +42,14 @@ namespace neam
     /// \brief Holds a progression entry
     struct duration_progression
     {
-      size_t timestamp; ///< \brief When the progression "has occured"
+      int64_t timestamp; ///< \brief When the progression "has occured"
       double value;     ///< \brief The value at the timestamp time
     };
 
     /// \brief Information about a measure point
     struct measure_point_entry
     {
-      size_t hit_count = 0; /// \brief Number of time the measure_point has been "hit"
+      uint64_t hit_count = 0; /// \brief Number of time the measure_point has been "hit"
       double value = 0;     /// \brief The average value
     };
 
@@ -59,24 +59,24 @@ namespace neam
       struct stack_entry
       {
 #ifdef _MSC_VER
-        stack_entry(size_t _se, size_t _si, size_t _csi, size_t _p) : self_index(_se), stack_index(_si), call_structure_index(_csi), parent(_p) {}
+        stack_entry(uint64_t _se, uint64_t _si, uint64_t _csi, uint64_t _p) : self_index(_se), stack_index(_si), call_structure_index(_csi), parent(_p) {}
 #endif
-        const size_t self_index; ///< \brief Hold the index of the stack_entry structure
-        const size_t stack_index; ///< \brief Hold the current stack index (index in the callgraph deque)
-        const size_t call_structure_index; ///< \brief Holds the index of the call struct
+        const uint64_t self_index; ///< \brief Hold the index of the stack_entry structure
+        const uint64_t stack_index; ///< \brief Hold the current stack index (index in the callgraph deque)
+        const uint64_t call_structure_index; ///< \brief Holds the index of the call struct
 
-        const size_t parent; ///< \brief Parent index
+        const uint64_t parent; ///< \brief Parent index
         // ----- //
-        std::vector<size_t> children = std::vector<size_t>(); ///< \brief Children indexes
+        std::vector<uint64_t> children = std::vector<uint64_t>(); ///< \brief Children indexes
 
-        size_t hit_count = 1; ///< \brief The number of time the entry has been hit
-        size_t fail_count = 0; ///< \brief The number of time that function failed
+        uint64_t hit_count = 1; ///< \brief The number of time the entry has been hit
+        uint64_t fail_count = 0; ///< \brief The number of time that function failed
 
         double average_self_time = 0; ///< \brief The average time consumed by the function and only this function
-        size_t average_self_time_count = 0; ///< \brief Number of time the self_time has been monitored
+        uint64_t average_self_time_count = 0; ///< \brief Number of time the self_time has been monitored
         std::deque<duration_progression> self_time_progression = std::deque<duration_progression>(); ///< \brief Hold the progression of the self_time average
         double average_global_time = 0; ///< \brief The average time consumed by the whole function call (including all its children)
-        size_t average_global_time_count = 0; ///< \brief Number of time the global_time has been monitored
+        uint64_t average_global_time_count = 0; ///< \brief Number of time the global_time has been monitored
         std::deque<duration_progression> global_time_progression = std::deque<duration_progression>();
 
         std::map<std::string, sequence> sequences = std::map<std::string, sequence>(); ///< \brief Hold sequences
@@ -94,19 +94,19 @@ namespace neam
         /// \brief push (or increment hit_count) into children a new call_info_struct.
         /// \note If not already in children, it will create a new stack_entry and insert its index at the end of children
         /// \return the reference of the stack_entry corresponding to the call_info_struct
-        stack_entry &push_children_call_info(size_t call_info_struct_index);
+        stack_entry &push_children_call_info(uint64_t call_info_struct_index);
 
         /// \brief Unlike push_children_call_info(), this method won't create anything / modify anithing
         /// It will simply lookup in the children array for a stack_entry for the call_info_struct_index call_info_struct
         /// If nothing is found, it returns nullptr
-        stack_entry *get_children_stack_entry(size_t call_info_struct_index) const;
+        stack_entry *get_children_stack_entry(uint64_t call_info_struct_index) const;
 
         /// \brief Returns a children stack_entry index corresponding to a given call_info_struct index
         /// \return -1 if nothing found
-        long get_children_stack_entry_index(size_t call_info_struct_index) const;
+        long get_children_stack_entry_index(uint64_t call_info_struct_index) const;
 
         /// \brief Start a stack
-        static stack_entry &initial_get_stack_entry(size_t call_info_struct_index);
+        static stack_entry &initial_get_stack_entry(uint64_t call_info_struct_index);
         /// \brief End a stack
         static void dispose_initial();
       };
